+++ docs/changelog.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 -   Add error message when user is not on a salesforce domain in the current tab
 -   Fix bug related to edit of entries. Now the changes are updated without the need of re-open the popup.
 -   Add toast messages to delete + edit entry
+-   If no entries are available for the current Org, the input form is shown directly. One click saved for the user.
 
 ## [0.4] - 26.02.2024